requirements.txt
alpaca-py==0.13.2
annotated-types==0.6.0
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
msgpack==1.0.7
numpy==1.26.2
pandas==2.1.3
peewee==3.17.0
pydantic==2.5.1
pydantic_core==2.14.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sseclient-py==1.8.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
websockets==11.0.3
